upgrade product/import product using ag-grid

+++ src/app/lib/custom-element/ag-list/column-define/command.define.ts
@@ -3,43 +3,48 @@ import { CommonService } from "../../../../services/common.service";
 import { AgButtonsCellRenderer } from "../cell/buttons.component";
 
 export const agMakeCommandColDef = (
-    cms: CommonService,
-    editAction: (data: any) => void,
-    deleteActin: (data: any) => void,
+  cms: CommonService,
+  editAction?: (data: any) => void,
+  deleteActin?: (data: any) => void,
 ): ColDef => {
-    return {
-        headerName: 'Sử/Xóa',
-        field: 'Command',
-        width: 110,
-        filter: false,
-        sortable: false,
-        pinned: 'right',
-        type: 'rightAligned',
-        cellClass: ['ag-cell-items-center', 'ag-cell-justify-center', 'ag-cell-no-padding-left', 'ag-cell-no-padding-right'],
-        cellRenderer: AgButtonsCellRenderer,
-        resizable: false,
-        cellStyle: { 'text-overflow': 'initial' },
-        cellRendererParams: {
-          buttons: [
-            {
-              name: 'edit',
-              status: 'warning',
-              icon: 'edit-2-outline',
-              outline: false,
-              action: (params: any, button: any) => {
-                editAction(params.node.data);
-              }
-            },
-            {
-              name: 'delete',
-              status: 'danger',
-              icon: 'trash-2-outline',
-              outline: false,
-              action: (params: any, button: any) => {
-                deleteActin(params.node.data);
-              }
-            },
-          ],
-        }
+  const buttons = [];
+  if (editAction) {
+    buttons.push({
+      name: 'edit',
+      status: 'warning',
+      icon: 'edit-2-outline',
+      outline: false,
+      action: (params: any, button: any) => {
+        editAction(params.node.data);
+      }
+    });
+  }
+  if (deleteActin) {
+    buttons.push({
+      name: 'delete',
+      status: 'danger',
+      icon: 'trash-2-outline',
+      outline: false,
+      action: (params: any, button: any) => {
+        deleteActin(params.node.data);
+      }
+    });
+  }
+  const config: ColDef = {
+    headerName: 'Sửa/Xóa',
+    field: 'Command',
+    width: 110,
+    filter: false,
+    sortable: false,
+    pinned: 'right',
+    type: 'rightAligned',
+    cellClass: ['ag-cell-items-center', 'ag-cell-justify-center', 'ag-cell-no-padding-left', 'ag-cell-no-padding-right'],
+    cellRenderer: AgButtonsCellRenderer,
+    resizable: false,
+    cellStyle: { 'text-overflow': 'initial' },
+    cellRendererParams: {
+      buttons: buttons,
     }
+  };
+  return config;
 }
